added optional timestamps
- timestamps may be added as an additional directory level.
- fix for misspelled method
- adaption of tests

+++ src/data/download_data.py
@@ -1,5 +1,6 @@
 import logging
 import os
+from datetime import datetime
 from os import path
 from typing import Callable, Optional
 import kagglehub as kh
@@ -7,7 +8,8 @@ import kagglehub as kh
 
 def download_data(
     raw_data_path: str,
-    download_func: Optional[Callable[[str], str]],
+    download_func: Optional[Callable[[], str]],
+    add_timestamp: bool = False,
 ) -> str:
     '''
     Download data, using the provided download function.
@@ -15,6 +17,7 @@ def download_data(
     Args:
         raw_data_path: Destination directory for the downloaded data
         download_func: Function to perform the download.
+        add_timestamp: Whether to add a timestamp directory level to raw_data_path
 
     Returns:
         Path to the downloaded dataset
@@ -31,10 +34,14 @@ def download_data(
     logging.info(f'>>> Downloading dataset ...')
     file_path = download_func()
 
-    # Ensure destination directory exists
-    os.makedirs(raw_data_path, exist_ok=True)
-
     # Move downloaded file to target directory
+    if not os.path.exists(file_path):
+        raise FileNotFoundError(f'Downloaded file not found: {file_path}')
+    if add_timestamp:
+        now = datetime.now()
+        timestamp = f'{now.strftime("%Y%m%d_%H%M")}'
+        raw_data_path = path.join(raw_data_path, timestamp)
+    os.makedirs(raw_data_path, exist_ok=True)
     new_file_path = path.join(raw_data_path, os.path.basename(file_path))
     os.rename(file_path, new_file_path)
 
@@ -46,13 +53,11 @@ def main() -> None:
     '''Main entry point for downloading data.'''
     def download_kaggle_data() -> str:
         dataset_id = 'ahmedlahlou/accidents-in-france-from-2005-to-2016'
-        return kh.datasets.download_dataset(dataset_id)
+        return kh.dataset_download(dataset_id)
 
-    root_dir = path.dirname(path.dirname(path.abspath(__file__)))
+    root_dir = path.join(path.dirname(path.abspath(__file__)), '..', '..')
     raw_data_path = path.join(root_dir, os.getenv('RAW_DATA_PATH', 'data/raw/'))
-    # Create the destination directory if it doesn't exist
-    os.makedirs(raw_data_path, exist_ok=True)
-    download_data(raw_data_path, download_kaggle_data)
+    download_data(raw_data_path, download_kaggle_data, add_timestamp=True)
 
 if __name__ == "__main__":
     main()
